% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_enviornmentSetup.R
\name{environmentSetup}
\alias{environmentSetup}
\title{GIMMEgVAR: Environment Setup}
\usage{
environmentSetup(outputPath)
}
\arguments{
\item{outputPath}{The user specified path to the directory where results files should be stored.}
}
\description{
Loads needed packages and creates directories
where logfiles and output from 'GIMMEgVAR'
will be stored
}
